An user should be able to update and delete their profile #49 (#55)
* Change db schema for profiles and add delete button

* Refactor

* Revert "Refactor"

This reverts commit eeac330d39781b7f06d416abb05c9bec8280404b.

* Revert "Revert "Refactor""

This reverts commit 23f737165549c6c97c29f1c8413d8be4fd793f10.

* Refactor classes; Create user delete modal component;

* WIP 

* Login cleanup of test code

* Cleanup of server.ts

* WIP: refactor profile image download to be from private bucket 

* Add functionality to load profile images from Blob

This commit includes the import of necessary libraries and context along with a useEffect hook to handle image data conversion. We convert the ArrayBuffer to a Blob and assign it to a URL, setting the source of the Avatar component to this URL to display the profile image.

* Upgrade @types/react and @types/react-dom versions

This update changes @types/react version from 18.2.14 to 18.2.43, and @types/react-dom version from 18.2.6 to 18.2.17. This upgrade aims to keep the project up-to-date with the latest changes and improvements in these packages.

* Add new ProfileProps and profile mapper functionality

Introduced ProfileProps in 'profile.prop.ts' specifying a new interface for profiles. Additionally, created 'profile.mapper.ts' that programmers can use to map Profile Entity to ProfileProps ensuring correct and safe type usage. The mapper also handles the default null values.

* Additional functionality for user profiles

Added functionality to upload and download user profile images in 'profile.server.ts'. Removed unnecessary models from 'schema.prisma'. Implemented image handling on the client side in 'EditProfile.tsx' and 'route.tsx'. Revamped error handling and made minor code style improvements across multiple files.

* Update action route and import statements

Updated the action route for the submit function in EditProfile.tsx.

* Cleanup of commented out code and uneeded imports

---------

Co-authored-by: Nicolas Allesøe <58513123+NicoCPH@users.noreply.github.com>
Co-authored-by: Cristian Florin Pandele <cfp@netcompany.com>